Update Footer to use Section component
