+++ src/Components/Footer/index.js
@@ -1,0 +1,2 @@
+export { default } from "./Footer"
+export * from "./Footer"
