+++ src/Components/Section/Section.jsx
@@ -1,6 +1,6 @@
 const Section = ({ customStyle, children }) => (
-    <section className={`w-full py-5 ${customStyle}`}>
-        <div className="lg:container mx-auto w-full px-3 flex flex-wrap items-center justify-between">
+    <section className={`w-full md:py-12 py-6 px-4 ${customStyle}`}>
+        <div className="lg:container mx-auto w-full flex flex-wrap items-center justify-between">
             { children }
         </div>
     </section>
