+++ src/app/about/page.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Footer from '../../Components/footer';
+import Footer from '../../Components/Footer';
 import Navbar from '../../Components/Navbar';
 import Image from 'next/image';
