+++ src/app/menu/page.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Navbar from '../../Components/Navbar';
-import Footer from '../../Components/footer';
+import Footer from '../../Components/Footer';
 import Image from 'next/image';
 
 const menuItems = [
